# openapi3.yaml
openapi: 3.0.1
info:
  title: 3d-extractable-management
  description: Manage extractable records with validation and audit logging
  version: 1.0.0
  license:
    name: 3d-extractable-management repo link
    url: https://github.com/MapColonies/3d-extractable-management/

security: []

tags:
  - name: records
    description: Record extractability management
  - name: audit_logs
    description: Documation of the the record history
  - name: users
    description: User authentication and validation

paths:
  /records:
    get:
      tags:
        - records
      summary: Get all extractable records
      operationId: getRecords
      responses:
        '200':
          description: List of extractable records (empty array if none found)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/extractable-record'
              examples:
                empty:
                  summary: No records
                  value: []
                withRecords:
                  summary: Some records
                  value:
                    - id: 123
                      record_name: rec_A
                      authorized_by: john_doe
                      data:
                        kuku: 'kuku random data'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                bad_request:
                  summary: Bad request
                  value:
                    isValid: false
                    message: 'Bad request – invalid query or parameters'
                    code: 'MISSING_CREDENTIALS'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

  /records/{record_name}:
    get:
      tags:
        - records
      summary: Get extractable record by record name
      operationId: getRecord
      parameters:
        - name: record_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/extractable-record'
              examples:
                found:
                  summary: Record exists
                  value:
                    id: 123
                    record_name: rec_A
                    authorized_by: john_doe
                    data:
                      kuku: 'kuku random data'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                not_found:
                  summary: Record does not exist
                  value:
                    isValid: false
                    message: "Record with name 'rec_X' not found"
                    code: 'INVALID_RECORD_NAME'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

    post:
      tags:
        - records
      summary: Create extractable record
      operationId: createRecord
      parameters:
        - name: record_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - authorized_by
              properties:
                username:
                  type: string
                password:
                  type: string
                authorized_by:
                  type: string
                  description: 'User authorizing this record creation'
                data:
                  type: object
                  additionalProperties: true
                  description: 'Optional metadata for the record'
      responses:
        '201':
          description: Record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/extractable-record'
              examples:
                created:
                  summary: Record created
                  value:
                    id: 123
                    record_name: rec_A
                    authorized_by: john_doe
                    data:
                      kuku: 'kuku random data'
        '400':
          description: Validation failed – call POST /records/validate first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                validation_failed:
                  summary: Validation missing
                  value:
                    isValid: false
                    message: 'Cannot create record – validation required'
                    code: 'MISSING_CREDENTIALS'
        '401':
          description: Unauthorized – invalid username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                unauthorized:
                  summary: Invalid credentials
                  value:
                    isValid: false
                    message: 'Invalid username or password'
                    code: 'INVALID_CREDENTIALS'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                record_not_found:
                  summary: Record does not exist
                  value:
                    isValid: false
                    message: "Record with name 'rec_X' not found"
                    code: 'INVALID_RECORD_NAME'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

    delete:
      tags:
        - records
      summary: Delete extractable record
      operationId: deleteRecord
      parameters:
        - name: record_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - authorized_by
              properties:
                username:
                  type: string
                password:
                  type: string
                authorized_by:
                  type: string
                  description: 'User authorizing this deletion'
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                not_found:
                  summary: Record missing
                  value:
                    isValid: false
                    message: "Cannot delete – record 'rec_X' not found"
                    code: 'INVALID_RECORD_NAME'
        '400':
          description: Validation failed – deletion not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                deletion_failed:
                  summary: Validation failed
                  value:
                    isValid: false
                    message: 'Deletion not allowed for this record'
                    code: 'MISSING_CREDENTIALS'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                unauthorized:
                  summary: Invalid credentials
                  value:
                    isValid: false
                    message: 'Invalid username or password'
                    code: 'INVALID_CREDENTIALS'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

  /records/validateCreate:
    post:
      tags:
        - records
      summary: Validate record can be created
      operationId: validateCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth-payload-with-record'
      responses:
        '200':
          description: Validation result for create
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                valid:
                  summary: Can create
                  value:
                    isValid: true
                    message: 'Record can be created'
                    code: 'SUCCESS'
        '401':
          description: Unauthorized – invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                invalid_credentials:
                  summary: Invalid credentials
                  value:
                    isValid: false
                    message: 'Invalid username or password'
                    code: 'INVALID_CREDENTIALS'
        '400':
          description: Validation failed – missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                missing_fields:
                  summary: Missing credentials or record_name
                  value:
                    isValid: false
                    message: 'Username, password, and record_name are required'
                    code: 'MISSING_CREDENTIALS'
        '404':
          description: Validation error – already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                missing_fields:
                  summary: Inappropriate record_name
                  value:
                    isValid: false
                    message: "Record 'rec_name' already exists"
                    code: 'INVALID_RECORD_NAME'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

  /records/validateDelete:
    post:
      tags:
        - records
      summary: Validate record can be deleted
      operationId: validateDelete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth-payload-with-record'
      responses:
        '200':
          description: Validation result for delete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                valid:
                  summary: Can delete
                  value:
                    isValid: true
                    message: 'Record can be deleted'
                    code: 'SUCCESS'
        '401':
          description: Unauthorized – invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                invalid_credentials:
                  summary: Invalid credentials
                  value:
                    isValid: false
                    message: 'Invalid username or password'
                    code: 'INVALID_CREDENTIALS'
        '400':
          description: Validation failed – missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                missing_fields:
                  summary: Missing credentials or record_name
                  value:
                    isValid: false
                    message: 'Username, password, and record_name are required'
                    code: 'MISSING_CREDENTIALS'
        '404':
          description: Validation error – doesn't exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                missing_fields:
                  summary: Inappropriate record_name
                  value:
                    isValid: false
                    message: "Record 'rec_name' doesn't exists"
                    code: 'INVALID_RECORD_NAME'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

  /audit/{record_name}:
    get:
      tags:
        - audit_logs
      summary: Get all audit logs by record name
      operationId: getAudit
      parameters:
        - name: record_name
          in: path
          required: true
          schema:
            type: string
          description: The record_name to fetch audit logs for
      responses:
        '200':
          description: List of audit logs for the record (empty array if none found)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/audit-log'
              examples:
                withAudits:
                  summary: Audit logs exist
                  value:
                    - id: 1
                      record_name: rec_A
                      authorized_by: john_doe
                      action: 'CREATE'
                    - id: 2
                      record_name: rec_A
                      authorized_by: jane_doe
                      action: 'DELETE'
                empty:
                  summary: No audit logs
                  value: []
        '400':
          description: Invalid record_name or request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                invalid_format:
                  summary: Invalid record_name
                  value:
                    isValid: false
                    message: 'Invalid record_name provided'
                    code: 'MISSING_CREDENTIALS'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

  /users/validate:
    post:
      tags:
        - users
      summary: Validate a user's username and password
      operationId: validateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth-payload'
      responses:
        '200':
          description: User credentials are valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                valid:
                  summary: Credentials valid
                  value:
                    isValid: true
                    message: 'User credentials are valid'
                    code: 'SUCCESS'
        '400':
          description: Missing username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                missing_fields:
                  summary: Missing credentials
                  value:
                    isValid: false
                    message: 'Username and password are required'
                    code: 'MISSING_CREDENTIALS'
        '401':
          description: Unauthorized – invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                invalid:
                  summary: Invalid credentials
                  value:
                    isValid: false
                    message: 'Invalid username or password'
                    code: 'INVALID_CREDENTIALS'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validateResponse'
              examples:
                internal_error:
                  summary: Unexpected server error
                  value:
                    isValid: false
                    message: 'Internal server error, please try again later'
                    code: 'INTERNAL_ERROR'

components:
  schemas:
    validateResponse:
      type: object
      required:
        - isValid
        - message
        - code
      properties:
        isValid:
          type: boolean
        message:
          type: string
        code:
          type: string
          enum:
            - SUCCESS
            - MISSING_CREDENTIALS
            - INVALID_CREDENTIALS
            - INVALID_RECORD_NAME
            - INTERNAL_ERROR

    auth-payload:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    auth-payload-with-record:
      type: object
      required:
        - username
        - password
        - record_name
      properties:
        username:
          type: string
        password:
          type: string
        record_name:
          type: string

    extractable-record:
      type: object
      required:
        - id
        - record_name
        - authorized_by
      properties:
        id:
          type: integer
          format: int64
        record_name:
          type: string
        authorized_by:
          type: string
        data:
          type: object
          additionalProperties: true
          description: Metadata stored in extractable_records.data

    audit-log:
      type: object
      required:
        - id
        - record_name
        - authorized_by
        - action
      properties:
        id:
          type: integer
          format: int64
        record_name:
          type: string
        authorized_by:
          type: string
        action:
          type: string
          enum:
            - CREATE
            - DELETE